Improved the Mac launcher to set the correct dock icon, and launch TB in JDK 8.
A couple of new icons were created, a 128x128 transparent png, and a mac-specific applet.icns version of that same transparent image.
The Mac launcher was improved to both show the new icon on it's on the launcher file, as well as show the new icon on the Mac dock while touchboard is running.
Note that the launcher exists specifically to run Touchboard in Java JDK version 8 (JDK 8). Later versions of Java currently have a bug on macOS that prevents them from typing capital letters properly.
See also: https://bugs.openjdk.org/browse/JDK-8302618